pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pancake-admin"
version = "0.1.0"
description = "A front-end and back-end separated permission management system based on FastAPI, JWT, Vue 3, and Naive UI."
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "HoHux", email = "jack@hohux.com"},
]
keywords = ["fastapi", "openapi", "mcp", "llm", "claude", "ai", "tools", "api", "dashboard", "admin-system"]
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.127.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "ruff>=0.14.10",
    "snowflake-id>=1.0.2",
    "sqlalchemy[asyncio]>=2.0.45",
]


[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    "alembic",
    ".git",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
addopts = "-v -s --strict-markers"

[tool.hatch.build.targets.wheel]
packages = ["app"]
